chore(cli): extract vite config as a submodule to the library
