+++ packages/cli/templates/js/tsconfig.json
@@ -6,8 +6,5 @@
     "resolveJsonModule": true,
     "allowSyntheticDefaultImports": true
   },
-  "include": [
-    "vite.config.ts",
-    "package.json"
-  ]
+  "include": ["vite.config.ts"]
 }
