+++ packages/field-plugin/helpers/vite/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "helper-vite",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "typescript": "^5.0.2",
+    "vite": "^4.4.5",
+    "vite-plugin-dts": "3.5.0"
+  }
+}
